Removes the toggle for switching between bounty range.
In Nov 2022 we updated our bounty award ranges and, to make calculating
bounties easier depending on when the report was submitted, we had a
toggle.

This MR removes that toggle, but still allows the old bounty range to
be shown if it is provided explicitly (&range=old). A warning icon is
also displayed.

The changes to remove the bounty range from the markdown and URL make
communication more efficient. It will also be straightforward to add
those back if/when we update our bounty amounts in the future.